+++ src/controllers/SettingsController.php
@@ -264,7 +264,8 @@ class SettingsController extends BaseSettingsController
         // Set the simple stuff
         $settings->defaultNotificationContactIds = $this->request->getBodyParam('defaultNotificationContactIds', $settings->defaultNotificationContactIds) ?: null;
         $settings->showSendoutTitleField = $this->request->getBodyParam('showSendoutTitleField', $settings->showSendoutTitleField) ?: false;
-        $settings->maxBatchSize = $this->request->getBodyParam('maxBatchSize', $settings->maxBatchSize) ?: null;
+        $settings->sendoutJobBatchSize = $this->request->getBodyParam('sendoutJobBatchSize', $settings->sendoutJobBatchSize) ?: null;
+        $settings->sendoutJobBatchDelay = $this->request->getBodyParam('sendoutJobBatchDelay', $settings->sendoutJobBatchDelay) ?: 0;
 
         // Save it
         if (!Craft::$app->getPlugins()->savePluginSettings(Campaign::$plugin, $settings->getAttributes())) {